moved crypto functions to wallet_util
